Move pkg/kwokctl/k8s to pkg/kwokctl/components
