--- pkg/kwokctl/k8s/kube_apiserver_tracing_config.yaml.tpl
@@ -1,4 +1,0 @@
-apiVersion: apiserver.config.k8s.io/v1alpha1
-kind: TracingConfiguration
-endpoint: {{ .Endpoint }}
-samplingRatePerMillion: 1000000
