+++ pkg/kwokctl/runtime/binary/cluster.go
@@ -32,7 +32,6 @@ import (
 	"sigs.k8s.io/kwok/pkg/consts"
 	"sigs.k8s.io/kwok/pkg/kwokctl/components"
 	"sigs.k8s.io/kwok/pkg/kwokctl/dryrun"
-	"sigs.k8s.io/kwok/pkg/kwokctl/k8s"
 	"sigs.k8s.io/kwok/pkg/kwokctl/runtime"
 	"sigs.k8s.io/kwok/pkg/log"
 	"sigs.k8s.io/kwok/pkg/utils/exec"
@@ -383,7 +382,7 @@ func (c *Cluster) addKubeApiserver(ctx context.Context, env *env) (err error) {
 
 	kubeApiserverTracingConfigPath := ""
 	if conf.JaegerOtlpGrpcPort != 0 {
-		kubeApiserverTracingConfigData, err := k8s.BuildKubeApiserverTracingConfig(k8s.BuildKubeApiserverTracingConfigParam{
+		kubeApiserverTracingConfigData, err := components.BuildKubeApiserverTracing(components.BuildKubeApiserverTracingConfig{
 			Endpoint: net.LocalAddress + ":" + format.String(conf.JaegerOtlpGrpcPort),
 		})
 		if err != nil {
