+++ pkg/kwokctl/runtime/compose/cluster.go
@@ -27,7 +27,6 @@ import (
 	"sigs.k8s.io/kwok/pkg/consts"
 	"sigs.k8s.io/kwok/pkg/kwokctl/components"
 	"sigs.k8s.io/kwok/pkg/kwokctl/dryrun"
-	"sigs.k8s.io/kwok/pkg/kwokctl/k8s"
 	"sigs.k8s.io/kwok/pkg/kwokctl/runtime"
 	"sigs.k8s.io/kwok/pkg/log"
 	"sigs.k8s.io/kwok/pkg/utils/exec"
@@ -424,7 +423,7 @@ func (c *Cluster) addKubeApiserver(ctx context.Context, env *env) (err error) {
 
 	kubeApiserverTracingConfigPath := ""
 	if conf.JaegerPort != 0 {
-		kubeApiserverTracingConfigData, err := k8s.BuildKubeApiserverTracingConfig(k8s.BuildKubeApiserverTracingConfigParam{
+		kubeApiserverTracingConfigData, err := components.BuildKubeApiserverTracing(components.BuildKubeApiserverTracingConfig{
 			Endpoint: c.Name() + "-jaeger:4317",
 		})
 		if err != nil {
