+++ test/e2e/kwokctl/dryrun/testdata/nerdctl/create_cluster_with_verbosity.txt
@@ -10,8 +10,8 @@ nerdctl pull registry.k8s.io/etcd:3.6.10-0
 nerdctl pull registry.k8s.io/kube-apiserver:v1.34.8
 cat <<EOF ><ROOT_DIR>/workdir/clusters/<CLUSTER_NAME>/apiserver-tracing-config.yaml
 apiVersion: apiserver.config.k8s.io/v1alpha1
-kind: TracingConfiguration
 endpoint: kwok-<CLUSTER_NAME>-jaeger:4317
+kind: TracingConfiguration
 samplingRatePerMillion: 1000000
 EOF
 nerdctl pull registry.k8s.io/kubectl:v1.34.8
